pyproject.toml
[project]
name = "note-gen"
version = "0.1.0"
description = "Offline lecture note generation with Whisper and LLM-powered study tools"
requires-python = ">=3.12,<3.14"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic-settings>=2.0",
    "python-dotenv>=1.0",
    "faster-whisper>=1.0",
    "sounddevice>=0.4",
    "soundfile>=0.12",
    "pymupdf>=1.24",
    "python-pptx>=1.0",
    "groq>=0.20",
    "python-multipart>=0.0.12",
    "numpy>=2.0",
    "aiofiles>=24.0",
    "aiosqlite>=0.20",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]
